Update UI and docs from alephclient to openaleph CLI
With OpenAlpeh the previous alephclient was renamed to openaleph as
command-line tool. This let to inconsistency. Also previous links to
OCCRP's docs were already broken (Error 404).

This commit reflects this change in the UI and docs.

Please note there are other mentions of alephclient that I did not
change.

$ grep -r alephclient                                                                                                                                                                                                               (change-docs-alephclient✱)
CHANGELOG.md:- Ability to map entities into lists via the UI and alephclient.
CHANGELOG.md:- Aleph has two new APIs for doing a collection `reingest` and `reindex`. The existing `process` collection API is gone. `alephclient` now supports running `reingest`, `reindex`, and `delete` on a collection.
CHANGELOG.md:Be advised that any data loaded via the entity mapping mechanism will need to be re-loaded after this. It is also worth noting that at OCCRP, we have now started generating mapped data via the `followthemoney` command-line tool, and are using `alephclient` to bulk-load the resulting stream of entities into the system. This has proven to be significantly quicker than the built-in mapping process.
aleph/logic/documents.py:        # to be consistent with the behaviour of alephclient
poetry.lock:aleph = ["alephclient (>=2.6.0,<3.0.0)", "furl (>=2.1.4,<3.0.0)"]